Extract body map utilities to body_map.py for reuse by result_handler
